ensure predictions are in float32 (and not float64), adjust masks from int8 to uint8, add to tests

+++ s2cloudless/cloud_detector.py
@@ -104,7 +104,7 @@ class S2PixelCloudDetector:
 
         proba = self.classifier.image_predict_proba(data, **kwargs)[..., 1]
 
-        return proba
+        return proba.astype(np.float32)
 
     def get_cloud_masks(self, data: np.ndarray, **kwargs: Any) -> np.ndarray:
         """
@@ -135,14 +135,14 @@ class S2PixelCloudDetector:
 
         if self.average_over:
             cloud_masks = np.asarray(
-                [convolve(cloud_prob, self.conv_filter) > threshold for cloud_prob in cloud_probs], dtype=np.int8
+                [convolve(cloud_prob, self.conv_filter) > threshold for cloud_prob in cloud_probs], dtype=np.uint8
             )
         else:
             cloud_masks = (cloud_probs > threshold).astype(np.int8)
 
         if self.dilation_size:
             cloud_masks = np.asarray(
-                [dilation(cloud_mask, self.dilation_filter) for cloud_mask in cloud_masks], dtype=np.int8
+                [dilation(cloud_mask, self.dilation_filter) for cloud_mask in cloud_masks], dtype=np.uint8
             )
 
         return cloud_masks
